Test: Use pythonz to provide the various versions of Python for tests
If pythonz is not found, this will exit with a complaint to avoid
unexpectedly doing massive amounts of Python building; use the -B
option to automatically install pythonz and build the versions we
need.

Unfortunately, it seems difficult to convince tox to use a Python
binary that's not in the path (see the comments in tox.ini for
details), so we work around this by prefixing all of the pythonz bin/
dirs to the path. Hopefully we can later find some way to do this more
cleanly.