Test: Uniquify Python versions so we don't try to any version twice
pythonz simply printed a message that the version was already installed
if we tried to install it a second time, but it's nicer (and marginally
faster) if we don't ask for any version to be installed a second time.

We also now sort the build order numerically by Python minor version
number.