tox.ini: skip_missing_interpreters = true
We want to fail when we can't test certain interpreters on a developer's
system.

The main change this produces is that the exit code is 255 instead of 0
when one or more interpreters are missing, and the list of testenvs at the
end has red FAIL messages for those instead of yellow SKIP messages.