pyproject.toml: Add 'pytest' as a dependency (doh!)
This was checked manually to see that it does not interfere with
selection of the correct version of pytest to install in the tox
environments. (We test against 5, 7 and 8.) But it might be a good
idea to tweak check-py-ver to test that the pytest version matches
the $TOX_ENV_NAME as well.