pyproject.toml: Add more metadata for PyPI page
We slightly reformat things, and add the license information,
project URLs and classifiers.

We set the required version of python to ≥3.6, even though we only
test on 3.7 and above because tox is unhappy with 3.6. If it's too
difficult to test it yourself, make your users do it for you. :-P