BROKEN Add mypy checking
This is broken in several ways, and needs a fair amount of work and
cleanup to get this happening.

It was inspired by pdoc in another project complaining about pytest_pt.py
with:

    Warn: Error parsing type annotation typing.Optional[ForwardRef('Module')] for psrc.pytest_pt.pt7_pytest_collect_file. Import of Module failed: name 'Module' is not defined (/home/cjs/co/public/gh/mc68-net/r8format/.build/virtualenv/lib/python3.12/site-packages/pdoc/doc_types.py:148)

That doesn't come up here, but looking at it, it appears we might actually
be using the wrong `Module` class (ast.Module instead of _pytest.Module or
_pytest.python.Module) for that declaration.

There will of course be further interesting issues given that this
file actually covers several different versions of pytest. In the end,
it may not be worth a "proper" fix for this, though perhaps we can still
run mypy as a side thing in various versions and see what happens.