Rename project: pytest_pt → pytest-pt (i.e., `_` → `-`)
This gives us consistency with other pytest plugins, such as
pytest-xdist, pytest-cache, and so on.

In fact, we were already called `pytest-pt` on pypi[1] because it
automatically converts underscores to hyphens in package names, and
installing with either name worked (and will continue to work) because
either is accepted in place of the other by Pip et al.

So this change is really just making the rest of the names (official
module name, our GitHub repo, the package files we upload, etc.)
consistent with this. (The GitHub repo has been renamed, but fetches
from the old name will continue to work, and old URLs will redirect to
the new one.)

[1]: https://pypi.org/project/pytest-pt/